REFACTOR: 상대경로를 절대경로로 통일
