+++ src/cards/api-connection/ApiConnectionCard/index.tsx
@@ -1,10 +1,9 @@
 import { Alert, Box, Stack } from "@mui/material";
 
 import Card from "@/cards/Card";
+import ExchangeButton from "@/cards/api-connection/ApiConnectionCard/ExchangeButton";
 import ExchangeSelectDescription from "@/components/signin/dialog/ExchangeSelectDescription";
 import { useUserExchangesQuery } from "@/hooks/query/useApiConnection";
-
-import ExchangeButton from "./ExchangeButton";
 
 interface Props {
   processing?: ExchangeType;
