+++ src/cards/dashboard/AdvertisementCard/index.tsx
@@ -3,9 +3,8 @@ import { useEffect, useRef, useState } from "react";
 import { Stack } from "@mui/material";
 
 import Card from "@/cards/Card";
-
-import AdvertisementIndex from "./AdvertisementIndex";
-import AdvertisementInfo from "./AdvertisementInfo";
+import AdvertisementIndex from "@/cards/dashboard/AdvertisementCard/AdvertisementIndex";
+import AdvertisementInfo from "@/cards/dashboard/AdvertisementCard/AdvertisementInfo";
 
 export default function AdvertisementCard() {
   const MAX_INDEX = 3;
