+++ src/cards/mypage/MypageCard/MypageExchange.tsx
@@ -1,14 +1,13 @@
 import { Typography } from "@mui/material";
 import { useAtom } from "jotai";
 
+import MypageInfo from "@/cards/mypage/MypageCard/MypageInfo";
 import ExchangeIcon from "@/components/Icon/ExchangeIcon";
 import SigninDialog from "@/components/dialog/SigninDialog";
 import useDialogGlobal from "@/components/dialog/useDialogGlobal";
 import ExchangeSelect from "@/components/signin/dialog/ExchangeSelect";
 import exchangeAtom from "@/datas/exchange";
 import { exchangeToKorean } from "@/libs/string";
-
-import MypageInfo from "./MypageInfo";
 
 export default function MypageExchange() {
   const [exchange, setExchange] = useAtom(exchangeAtom);
