+++ src/hooks/card/useCurrencySelect.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 
 import { UnfoldMore } from "@mui/icons-material";
 
-import CardHeaderSelect from "./CardHeaderSelect";
+import CardHeaderSelect from "@/hooks/card/CardHeaderSelect";
 
 function useCurrencySelect() {
   const [currency, setCurrency] = useState<string>("won");
