+++ src/hooks/card/useOrderSelect.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 
 import { UnfoldMore } from "@mui/icons-material";
 
-import CardHeaderSelect from "./CardHeaderSelect";
+import CardHeaderSelect from "@/hooks/card/CardHeaderSelect";
 
 function useOrderSelect() {
   const [order, setOrder] = useState<string>("latest");
